Adding initial ClipList component | Updating overall styles
A ClipList component has been added which initially displays the state for an empty list(containing no clips/items). The animation for the add button has been updated to rotate instead of scale. A copy icon has been added as the clipboard app icon in the header and several other small tweaks and updates have been perfomed to update the look of the overall popup display