Adding an initial version of the clip component
A clip component has been added and styled with some dummy content. The clips if present are populated in the cliplist and displayed to the user instead of the empty clip list (with pencil icon). The next step would be to style the scrollbar and add custom styling to match our project styles