Saving data to local storage
This commit contains the functionality to persist data to local storage. We're using the HTML5 local storage API to persist data to and retrieve it from the local storage. This was done as the useStorage hook provided by plasmo seems to wipe the data from the local storage when the extension popup is closed and reopened. In future we may want to replace the use of the local storage HTML5 API with the useStorage hook instead.