Disabling the delete button when no clip has been selected
The delete button gets disabled when the user opens the add clip modal to add a new clip since no clip has been selected to be deleted. The color of the button is changed to a lighter grey version and all hover and active state styles have been reverted so as to indicate the button is non interactive at this stage.